app/requirements.txt
flask
python-dotenv
flask-expects-json
redis
pika
